split both implementations up into separate modules
