+++ cranelift/filetests/src/test_wasm/env.rs
@@ -742,41 +742,6 @@ impl<'a> FuncEnvironment for FuncEnv<'a> {
         self.inner.translate_suspend(builder, tag_index)
     }
 
-    /// TODO(dhil): write documentation.
-    fn translate_cont_new_baseline(
-        &mut self,
-        builder: &mut cranelift_frontend::FunctionBuilder,
-        state: &cranelift_wasm::FuncTranslationState,
-        func: ir::Value,
-        arg_types: &[WasmValType],
-        return_types: &[WasmValType],
-    ) -> cranelift_wasm::WasmResult<ir::Value> {
-        self.inner
-            .translate_cont_new_baseline(builder, state, func, arg_types, return_types)
-    }
-
-    /// TODO(frank-emrich): write documentation.
-    fn translate_resume_baseline(
-        &mut self,
-        builder: &mut cranelift_frontend::FunctionBuilder,
-        type_index: u32,
-        contref: ir::Value,
-        resume_args: &[ir::Value],
-        resumetable: &[(u32, ir::Block)],
-    ) -> Vec<ir::Value> {
-        self.inner
-            .translate_resume_baseline(builder, type_index, contref, resume_args, resumetable)
-    }
-
-    /// TODO(dhil): write documentation.
-    fn translate_suspend_baseline(
-        &mut self,
-        builder: &mut cranelift_frontend::FunctionBuilder,
-        tag_index: ir::Value,
-    ) -> ir::Value {
-        self.inner.translate_suspend_baseline(builder, tag_index)
-    }
-
     /// TODO
     fn continuation_arguments(&self, type_index: u32) -> &[WasmValType] {
         self.inner.continuation_arguments(type_index)
@@ -785,16 +750,6 @@ impl<'a> FuncEnvironment for FuncEnv<'a> {
     /// TODO
     fn continuation_returns(&self, type_index: u32) -> &[WasmValType] {
         self.inner.continuation_returns(type_index)
-    }
-
-    /// TODO
-    fn typed_continuations_load_payloads(
-        &mut self,
-        builder: &mut cranelift_frontend::FunctionBuilder,
-        valtypes: &[ir::Type],
-    ) -> std::vec::Vec<ir::Value> {
-        self.inner
-            .typed_continuations_load_payloads(builder, valtypes)
     }
 
     /// TODO
@@ -825,34 +780,6 @@ impl<'a> FuncEnvironment for FuncEnv<'a> {
     ) -> ir::Value {
         self.inner
             .typed_continuations_load_continuation_object(builder)
-    }
-
-    fn typed_continuations_load_parent(
-        &mut self,
-        builder: &mut cranelift_frontend::FunctionBuilder,
-        contobj: ir::Value,
-    ) -> ir::Value {
-        self.inner.typed_continuations_load_parent(builder, contobj)
-    }
-
-    fn typed_continuations_store_parent(
-        &mut self,
-        builder: &mut cranelift_frontend::FunctionBuilder,
-        contobj: ir::Value,
-        new_parent: ir::Value,
-    ) {
-        self.inner
-            .typed_continuations_store_parent(builder, contobj, new_parent);
-    }
-
-    fn typed_continuations_load_return_values(
-        &mut self,
-        builder: &mut cranelift_frontend::FunctionBuilder,
-        valtypes: &[WasmValType],
-        contobj: ir::Value,
-    ) -> std::vec::Vec<ir::Value> {
-        self.inner
-            .typed_continuations_load_return_values(builder, valtypes, contobj)
     }
 
     fn typed_continuations_store_resume_args(
@@ -890,16 +817,6 @@ impl<'a> FuncEnvironment for FuncEnv<'a> {
             .typed_continuations_cont_ref_get_cont_obj(builder, contref)
     }
 
-    /// TODO
-    fn typed_continuations_drop_cont_obj(
-        &mut self,
-        builder: &mut cranelift_frontend::FunctionBuilder,
-        contobj: ir::Value,
-    ) {
-        self.inner
-            .typed_continuations_drop_cont_obj(builder, contobj)
-    }
-
     fn typed_continuations_load_tag_return_values(
         &mut self,
         builder: &mut cranelift_frontend::FunctionBuilder,
@@ -909,19 +826,5 @@ impl<'a> FuncEnvironment for FuncEnv<'a> {
         return self
             .inner
             .typed_continuations_load_tag_return_values(builder, contobj, valtypes);
-    }
-
-    /// TODO
-    fn typed_continuations_forward_tag_return_values(
-        &mut self,
-        builder: &mut cranelift_frontend::FunctionBuilder,
-        parent_contobj: ir::Value,
-        child_contobj: ir::Value,
-    ) {
-        return self.inner.typed_continuations_forward_tag_return_values(
-            builder,
-            parent_contobj,
-            child_contobj,
-        );
     }
 }
