+++ crates/cranelift/src/func_environ.rs
@@ -10,8 +10,8 @@ use cranelift_codegen::ir::{
 };
 use cranelift_codegen::isa::{self, CallConv, TargetFrontendConfig, TargetIsa};
 use cranelift_entity::{EntityRef, PrimaryMap};
+use cranelift_frontend::FunctionBuilder;
 use cranelift_frontend::Variable;
-use cranelift_frontend::{FunctionBuilder, Switch};
 use cranelift_wasm::{
     self, FuncIndex, FuncTranslationState, GlobalIndex, GlobalVariable, Heap, HeapData, HeapStyle,
     MemoryIndex, TableIndex, TagIndex, TargetEnvironment, TypeIndex, WasmHeapType, WasmRefType,
@@ -28,6 +28,12 @@ use wasmtime_environ::{
 };
 use wasmtime_environ::{FUNCREF_INIT_BIT, FUNCREF_MASK};
 
+#[cfg(feature = "typed_continuations_baseline_implementation")]
+use crate::wasmfx::baseline as wasmfx_impl;
+
+#[cfg(not(feature = "typed_continuations_baseline_implementation"))]
+use crate::wasmfx::optimized as wasmfx_impl;
+
 macro_rules! declare_function_signatures {
     (
         $(
@@ -37,7 +43,8 @@ macro_rules! declare_function_signatures {
     ) => {
         /// A struct with an `Option<ir::SigRef>` member for every builtin
         /// function, to de-duplicate constructing/getting its signature.
-        struct BuiltinFunctionSignatures {
+        /// NOTE(frank-emrich) pub for use in crate::wasmfx::* modules
+        pub (crate) struct BuiltinFunctionSignatures {
             pointer_type: ir::Type,
             reference_type: ir::Type,
             call_conv: isa::CallConv,
@@ -92,7 +99,8 @@ macro_rules! declare_function_signatures {
             }
 
             $(
-                fn $name(&mut self, func: &mut Function) -> ir::SigRef {
+                // NOTE(frank-emrich) pub for use in crate::wasmfx::* modules
+                pub (crate) fn $name(&mut self, func: &mut Function) -> ir::SigRef {
                     let sig = self.$name.unwrap_or_else(|| {
                         func.import_signature(Signature {
                             params: vec![ $( self.$param() ),* ],
@@ -112,7 +120,8 @@ wasmtime_environ::foreach_builtin_function!(declare_function_signatures);
 
 /// The `FuncEnvironment` implementation for use by the `ModuleEnvironment`.
 pub struct FuncEnvironment<'module_environment> {
-    isa: &'module_environment (dyn TargetIsa + 'module_environment),
+    /// NOTE(frank-emrich) pub for use in crate::wasmfx::* modules
+    pub(crate) isa: &'module_environment (dyn TargetIsa + 'module_environment),
     module: &'module_environment Module,
     types: &'module_environment ModuleTypesBuilder,
 
@@ -129,7 +138,8 @@ pub struct FuncEnvironment<'module_environment> {
     pcc_vmctx_memtype: Option<ir::MemoryType>,
 
     /// Caches of signatures for builtin functions.
-    builtin_function_signatures: BuiltinFunctionSignatures,
+    /// NOTE(frank-emrich) pub for use in crate::wasmfx::* modules
+    pub(crate) builtin_function_signatures: BuiltinFunctionSignatures,
 
     /// Offsets to struct fields accessed by JIT code.
     pub(crate) offsets: VMOffsets<u8>,
@@ -166,745 +176,6 @@ pub struct FuncEnvironment<'module_environment> {
     #[cfg(feature = "wmemcheck")]
     wmemcheck: bool,
 }
-
-#[macro_use]
-mod typed_continuation_helpers {
-    use super::IntCC;
-    use cranelift_codegen::ir;
-    use cranelift_codegen::ir::types::*;
-    use cranelift_codegen::ir::InstBuilder;
-    use cranelift_frontend::FunctionBuilder;
-    use std::mem;
-    use wasmtime_environ::BuiltinFunctionIndex;
-
-    // This is a reference to this very module.
-    // We need it so that we can refer to the functions inside this module from
-    // macros, such that the same path works when the macro is expanded inside
-    // or outside of this module.
-    use crate::func_environ::typed_continuation_helpers as tc;
-
-    /// Low-level implementation of debug printing. Do not use directly; see
-    /// `emit_debug_println!` macro for doing actual printing.
-    ///
-    /// Takes a string literal which may contain placeholders similarly to those
-    /// supported by `std::fmt`.
-    ///
-    /// Currently supported placeholders:
-    /// {}       for unsigned integers
-    /// {:p}     for printing pointers (in hex form)
-    ///
-    /// When printing, we replace them with the corresponding values in `vals`.
-    /// Thus, the number of placeholders in `s` must match the number of entries
-    /// in `vals`.
-    pub fn emit_debug_print<'a>(
-        env: &mut crate::func_environ::FuncEnvironment<'a>,
-        builder: &mut FunctionBuilder,
-        s: &'static str,
-        vals: &[ir::Value],
-    ) {
-        let print_s_infix = |env: &mut crate::func_environ::FuncEnvironment<'a>,
-                             builder: &mut FunctionBuilder,
-                             start: usize,
-                             end: usize| {
-            if start < end {
-                let s: &'static str = &s[start..end];
-                // This is quite dodgy, which is why we can only do this for
-                // debugging purposes:
-                // At jit time, we take a pointer to the slice of the (static)
-                // string, thus yielding an address within wasmtime's DATA
-                // section. This pointer is hard-code into generated code. We do
-                // not emit any kind of relocation information, which means that
-                // this breaks if we were to store the generated code and use it
-                // during subsequent executions of wasmtime (e.g., when using
-                // wasmtime compile).
-                let ptr = s.as_ptr();
-                let ptr = builder.ins().iconst(env.pointer_type(), ptr as i64);
-                let len = s.len();
-                let len = builder.ins().iconst(I64, len as i64);
-
-                let index = BuiltinFunctionIndex::tc_print_str();
-                let sig = env.builtin_function_signatures.tc_print_str(builder.func);
-                env.generate_builtin_call_no_return_val(builder, index, sig, vec![ptr, len]);
-            }
-        };
-        let print_int = |env: &mut crate::func_environ::FuncEnvironment<'a>,
-                         builder: &mut FunctionBuilder,
-                         val: ir::Value| {
-            let index = BuiltinFunctionIndex::tc_print_int();
-            let sig = env.builtin_function_signatures.tc_print_int(builder.func);
-            let ty = builder.func.dfg.value_type(val);
-            let val = match ty {
-                I32 => builder.ins().uextend(I64, val),
-                I64 => val,
-                _ => panic!("Cannot print type {}", ty),
-            };
-            env.generate_builtin_call_no_return_val(builder, index, sig, vec![val]);
-        };
-        let print_pointer = |env: &mut crate::func_environ::FuncEnvironment<'a>,
-                             builder: &mut FunctionBuilder,
-                             ptr: ir::Value| {
-            let index = BuiltinFunctionIndex::tc_print_pointer();
-            let sig = env
-                .builtin_function_signatures
-                .tc_print_pointer(builder.func);
-            env.generate_builtin_call_no_return_val(builder, index, sig, vec![ptr]);
-        };
-
-        if wasmtime_continuations::ENABLE_DEBUG_PRINTING {
-            let mut prev_end = 0;
-            let mut i = 0;
-
-            let mut ph_matches: Vec<(usize, &'static str)> = s
-                .match_indices("{}")
-                .chain(s.match_indices("{:p}"))
-                .collect();
-            ph_matches.sort_by_key(|(index, _)| *index);
-
-            for (start, matched_ph) in ph_matches {
-                let end = start + matched_ph.len();
-
-                assert!(
-                    i < vals.len(),
-                    "Must supply as many entries in vals as there are placeholders in the string"
-                );
-
-                print_s_infix(env, builder, prev_end, start);
-                match matched_ph {
-                    "{}" => print_int(env, builder, vals[i]),
-                    "{:p}" => print_pointer(env, builder, vals[i]),
-                    u => panic!("Unsupported placeholder in debug_print input string: {}", u),
-                }
-                prev_end = end;
-                i += 1;
-            }
-            assert_eq!(
-                i,
-                vals.len(),
-                "Must supply as many entries in vals as there are placeholders in the string"
-            );
-
-            print_s_infix(env, builder, prev_end, s.len());
-        }
-    }
-
-    /// Emits code to print debug information. Only actually prints in debug
-    /// builds and if debug printing flag is enabled. The third and all
-    /// following arguments are like those to println!: A string literal with
-    /// placeholders followed by the actual values.
-    ///
-    /// Summary of arguments:
-    /// * `env` - Type &mut crate::func_environ::FuncEnvironment<'a>
-    /// * `builder` - Type &mut FunctionBuilder,
-    /// * `msg` : String literal, containing placeholders like those supported by println!
-    /// * remaining arguments: ir::Values filled into the placeholders in `msg`
-    #[allow(unused_macros)]
-    macro_rules! emit_debug_println {
-        ($env : expr, $builder : expr, $msg : literal, $( $arg:expr ),*) => {
-            let msg_newline : &'static str= std::concat!(
-                $msg,
-                "\n"
-            );
-            tc::emit_debug_print($env, $builder, msg_newline, &[$($arg),*]);
-        }
-    }
-
-    /// Low-level implementation of assertion mechanism. Use emit_debug_* macros
-    /// instead.
-    ///
-    /// If `ENABLE_DEBUG_PRINTING` is enabled, `error_str` is printed before
-    /// trapping in case of an assertion violation.
-    pub fn emit_debug_assert_generic<'a>(
-        env: &mut crate::func_environ::FuncEnvironment<'a>,
-        builder: &mut FunctionBuilder,
-        condition: ir::Value,
-        error_str: &'static str,
-    ) {
-        if cfg!(debug_assertions) {
-            if wasmtime_continuations::ENABLE_DEBUG_PRINTING {
-                let failure_block = builder.create_block();
-                let continue_block = builder.create_block();
-
-                builder
-                    .ins()
-                    .brif(condition, continue_block, &[], failure_block, &[]);
-
-                builder.switch_to_block(failure_block);
-                builder.seal_block(failure_block);
-
-                emit_debug_print(env, builder, error_str, &[]);
-                builder.ins().debugtrap();
-                builder.ins().jump(continue_block, &[]);
-
-                builder.switch_to_block(continue_block);
-                builder.seal_block(continue_block);
-            } else {
-                builder
-                    .ins()
-                    .trapz(condition, ir::TrapCode::User(crate::DEBUG_ASSERT_TRAP_CODE));
-            }
-        }
-    }
-
-    /// Low-level implementation of assertion mechanism. Use emit_debug_* macros
-    /// instead.
-    ///
-    /// If `ENABLE_DEBUG_PRINTING` is enabled, `error_str` is printed before
-    /// trapping in case of an assertion violation. Here, `error_str` is expected
-    /// to contain two placeholders, such as {} or {:p}, which are replaced with
-    /// `v1` and `v2` when printing.
-    pub fn emit_debug_assert_icmp<'a>(
-        env: &mut crate::func_environ::FuncEnvironment<'a>,
-        builder: &mut FunctionBuilder,
-        operator: IntCC,
-        v1: ir::Value,
-        v2: ir::Value,
-        error_str: &'static str,
-    ) {
-        if cfg!(debug_assertions) {
-            let cmp_res = builder.ins().icmp(operator, v1, v2);
-
-            if wasmtime_continuations::ENABLE_DEBUG_PRINTING {
-                let failure_block = builder.create_block();
-                let continue_block = builder.create_block();
-
-                builder
-                    .ins()
-                    .brif(cmp_res, continue_block, &[], failure_block, &[]);
-
-                builder.switch_to_block(failure_block);
-                builder.seal_block(failure_block);
-
-                emit_debug_print(env, builder, error_str, &[v1, v2]);
-                builder.ins().debugtrap();
-                builder.ins().jump(continue_block, &[]);
-
-                builder.switch_to_block(continue_block);
-                builder.seal_block(continue_block);
-            } else {
-                builder
-                    .ins()
-                    .trapz(cmp_res, ir::TrapCode::User(crate::DEBUG_ASSERT_TRAP_CODE));
-            }
-        }
-    }
-
-    /// Used to implement other macros, do not use directly.
-    macro_rules! emit_debug_assert_icmp {
-        ( $env : expr,
-            $builder: expr,
-        $operator : expr,
-        $operator_string  : expr,
-        $v1 : expr,
-        $v2 : expr) => {
-            let msg: &'static str = std::concat!(
-                "assertion failure in ",
-                std::file!(),
-                ", line ",
-                std::line!(),
-                ": {} ",
-                $operator_string,
-                " {} does not hold\n"
-            );
-            tc::emit_debug_assert_icmp($env, $builder, $operator, $v1, $v2, msg);
-        };
-    }
-
-    macro_rules! emit_debug_assert {
-        ($env: expr, $builder: expr, $condition: expr) => {
-            let msg: &'static str = std::concat!(
-                "assertion failure in ",
-                std::file!(),
-                ", line ",
-                std::line!(),
-                "\n"
-            );
-            tc::emit_debug_assert_generic($env, $builder, $condition, msg);
-        };
-    }
-
-    macro_rules! emit_debug_assert_eq {
-        ($env: expr, $builder: expr, $v1 : expr, $v2: expr) => {
-            emit_debug_assert_icmp!($env, $builder, IntCC::Equal, "==", $v1, $v2);
-        };
-    }
-
-    macro_rules! emit_debug_assert_ne {
-        ($env: expr, $builder: expr, $v1 : expr, $v2: expr) => {
-            emit_debug_assert_icmp!($env, $builder, IntCC::NotEqual, "!=", $v1, $v2);
-        };
-    }
-
-    macro_rules! emit_debug_assert_ule {
-        ($env: expr, $builder: expr, $v1 : expr, $v2: expr) => {
-            emit_debug_assert_icmp!(
-                $env,
-                $builder,
-                IntCC::UnsignedLessThanOrEqual,
-                "<=",
-                $v1,
-                $v2
-            );
-        };
-    }
-
-    #[derive(Copy, Clone)]
-    pub struct ContinuationObject {
-        address: ir::Value,
-        pointer_type: ir::Type,
-    }
-
-    #[derive(Copy, Clone)]
-    pub struct Payloads {
-        /// Base address of this object, which must be shifted by `offset` below.
-        base: ir::Value,
-
-        /// Adding this (statically) known offset gets us the overall address.
-        offset: i32,
-
-        pointer_type: ir::Type,
-    }
-
-    #[derive(Copy, Clone)]
-    pub struct VMContext {
-        address: ir::Value,
-        pointer_type: ir::Type,
-    }
-
-    impl ContinuationObject {
-        pub fn new(address: ir::Value, pointer_type: ir::Type) -> ContinuationObject {
-            ContinuationObject {
-                address,
-                pointer_type,
-            }
-        }
-
-        pub fn args(&self) -> Payloads {
-            let offset = wasmtime_continuations::offsets::continuation_object::ARGS;
-            Payloads::new(self.address, offset, self.pointer_type)
-        }
-
-        pub fn tag_return_values(&self) -> Payloads {
-            let offset = wasmtime_continuations::offsets::continuation_object::TAG_RETURN_VALUES;
-            Payloads::new(self.address, offset, self.pointer_type)
-        }
-
-        /// Loads the value of the `state` field of the continuation object,
-        /// which is represented using the `State` enum.
-        fn load_state(&self, builder: &mut FunctionBuilder) -> ir::Value {
-            let mem_flags = ir::MemFlags::trusted();
-            let offset = wasmtime_continuations::offsets::continuation_object::STATE;
-
-            // Let's make sure that we still represent the State enum as i32.
-            debug_assert!(mem::size_of::<wasmtime_continuations::State>() == mem::size_of::<i32>());
-
-            builder.ins().load(I32, mem_flags, self.address, offset)
-        }
-
-        /// Sets the value of the `state` field of the continuation object,
-        pub fn set_state(
-            &self,
-            builder: &mut FunctionBuilder,
-            state: wasmtime_continuations::State,
-        ) {
-            let mem_flags = ir::MemFlags::trusted();
-            let offset = wasmtime_continuations::offsets::continuation_object::STATE;
-
-            // Let's make sure that we still represent the State enum as i32.
-            debug_assert!(mem::size_of::<wasmtime_continuations::State>() == mem::size_of::<i32>());
-
-            let v = builder.ins().iconst(I32, state.discriminant() as i64);
-            builder.ins().store(mem_flags, v, self.address, offset);
-        }
-
-        /// Checks whether the continuation object is invoked (i.e., `resume`
-        /// was called at least once on the object).
-        pub fn is_invoked(&self, builder: &mut FunctionBuilder) -> ir::Value {
-            // TODO(frank-emrich) In the future, we may get rid of the State field
-            // in `ContinuationObject` and try to infer the state by other means.
-            // For example, we may alllocate the `ContinuationFiber` lazily, doing
-            // so only at the point when a continuation is actualy invoked, meaning
-            // that we can use the null-ness of the `fiber` field as an indicator
-            // for invokedness.
-            let actual_state = self.load_state(builder);
-            let invoked: i32 = i32::from(wasmtime_continuations::State::Invoked);
-            builder
-                .ins()
-                .icmp_imm(IntCC::Equal, actual_state, invoked as i64)
-        }
-
-        /// Checks whether the continuation object has returned (i.e., the
-        /// function used as continuation has returned normally).
-        pub fn has_returned(&self, builder: &mut FunctionBuilder) -> ir::Value {
-            let actual_state = self.load_state(builder);
-            let returned: i32 = i32::from(wasmtime_continuations::State::Returned);
-            builder
-                .ins()
-                .icmp_imm(IntCC::Equal, actual_state, returned as i64)
-        }
-
-        /// Returns pointer to buffer where results are stored after a
-        /// continuation has returned.
-        pub fn get_results<'a>(
-            &self,
-            env: &mut crate::func_environ::FuncEnvironment<'a>,
-            builder: &mut FunctionBuilder,
-        ) -> ir::Value {
-            if cfg!(debug_assertions) {
-                let has_returned = self.has_returned(builder);
-                emit_debug_assert!(env, builder, has_returned);
-            }
-            return self.args().get_data(builder);
-        }
-    }
-
-    impl Payloads {
-        pub(crate) fn new(base: ir::Value, offset: i32, pointer_type: ir::Type) -> Payloads {
-            Payloads {
-                base,
-                offset,
-                pointer_type,
-            }
-        }
-
-        fn get(&self, builder: &mut FunctionBuilder, ty: ir::Type, offset: i32) -> ir::Value {
-            let mem_flags = ir::MemFlags::trusted();
-            builder
-                .ins()
-                .load(ty, mem_flags, self.base, self.offset + offset)
-        }
-
-        fn set<T>(&self, builder: &mut FunctionBuilder, offset: i32, value: ir::Value) {
-            debug_assert_eq!(
-                builder.func.dfg.value_type(value),
-                Type::int_with_byte_size(std::mem::size_of::<T>() as u16).unwrap()
-            );
-            let mem_flags = ir::MemFlags::trusted();
-            builder
-                .ins()
-                .store(mem_flags, value, self.base, self.offset + offset);
-        }
-
-        pub fn get_data(&self, builder: &mut FunctionBuilder) -> ir::Value {
-            self.get(
-                builder,
-                self.pointer_type,
-                wasmtime_continuations::offsets::payloads::DATA,
-            )
-        }
-
-        fn get_capacity(&self, builder: &mut FunctionBuilder) -> ir::Value {
-            let ty = Type::int_with_byte_size(std::mem::size_of::<
-                wasmtime_continuations::types::payloads::Capacity,
-            >() as u16)
-            .unwrap();
-            self.get(
-                builder,
-                ty,
-                wasmtime_continuations::offsets::payloads::CAPACITY,
-            )
-        }
-
-        fn get_length(&self, builder: &mut FunctionBuilder) -> ir::Value {
-            let ty = Type::int_with_byte_size(std::mem::size_of::<
-                wasmtime_continuations::types::payloads::Length,
-            >() as u16)
-            .unwrap();
-            self.get(
-                builder,
-                ty,
-                wasmtime_continuations::offsets::payloads::LENGTH,
-            )
-        }
-
-        fn set_length(&self, builder: &mut FunctionBuilder, length: ir::Value) {
-            self.set::<wasmtime_continuations::types::payloads::Length>(
-                builder,
-                wasmtime_continuations::offsets::payloads::LENGTH,
-                length,
-            );
-        }
-
-        fn set_capacity(&self, builder: &mut FunctionBuilder, capacity: ir::Value) {
-            self.set::<wasmtime_continuations::types::payloads::Capacity>(
-                builder,
-                wasmtime_continuations::offsets::payloads::CAPACITY,
-                capacity,
-            );
-        }
-
-        fn set_data(&self, builder: &mut FunctionBuilder, data: ir::Value) {
-            self.set::<*mut u8>(
-                builder,
-                wasmtime_continuations::offsets::payloads::DATA,
-                data,
-            );
-        }
-
-        /// Returns pointer to next empty slot in data buffer and marks the
-        /// subsequent `arg_count` slots as occupied.
-        pub fn occupy_next_slots<'a>(
-            &self,
-            env: &mut crate::func_environ::FuncEnvironment<'a>,
-            builder: &mut FunctionBuilder,
-            arg_count: i32,
-        ) -> ir::Value {
-            let data = self.get_data(builder);
-            let original_length = self.get_length(builder);
-            let new_length = builder.ins().iadd_imm(original_length, arg_count as i64);
-            self.set_length(builder, new_length);
-
-            if cfg!(debug_assertions) {
-                let capacity = self.get_capacity(builder);
-                emit_debug_assert_ule!(env, builder, new_length, capacity);
-            }
-
-            let value_size =
-                mem::size_of::<wasmtime_continuations::types::payloads::DataEntries>() as i64;
-            let byte_offset = builder.ins().imul_imm(original_length, value_size);
-            builder.ins().iadd(data, byte_offset)
-        }
-
-        #[allow(dead_code)]
-        pub fn deallocate_buffer<'a>(
-            &self,
-            env: &mut crate::func_environ::FuncEnvironment<'a>,
-            builder: &mut FunctionBuilder,
-        ) {
-            let zero = builder.ins().iconst(ir::types::I64, 0);
-            let capacity = self.get_capacity(builder);
-            emit_debug_assert_ne!(env, builder, capacity, zero);
-
-            let align = builder.ins().iconst(
-                I64,
-                std::mem::align_of::<wasmtime_continuations::types::payloads::DataEntries>() as i64,
-            );
-            let entry_size =
-                std::mem::size_of::<wasmtime_continuations::types::payloads::DataEntries>();
-            let size = builder.ins().imul_imm(capacity, entry_size as i64);
-
-            let index = BuiltinFunctionIndex::tc_deallocate();
-            let sig = env.builtin_function_signatures.tc_deallocate(builder.func);
-
-            let ptr = self.get_data(builder);
-            env.generate_builtin_call_no_return_val(builder, index, sig, vec![ptr, size, align]);
-
-            self.set_capacity(builder, zero);
-            self.set_length(builder, zero);
-            self.set_data(builder, zero);
-        }
-
-        pub fn ensure_capacity<'a>(
-            &self,
-            env: &mut crate::func_environ::FuncEnvironment<'a>,
-            builder: &mut FunctionBuilder,
-            required_capacity: ir::Value,
-        ) {
-            let zero = builder.ins().iconst(ir::types::I64, 0);
-            emit_debug_assert_ne!(env, builder, required_capacity, zero);
-
-            if cfg!(debug_assertions) {
-                let data = self.get_data(builder);
-                emit_debug_println!(
-                    env,
-                    builder,
-                    "[ensure_capacity] contobj/base {:p}, buffer is {:p}",
-                    self.base,
-                    data
-                );
-            }
-
-            let capacity = self.get_capacity(builder);
-
-            let sufficient_capacity_block = builder.create_block();
-            let insufficient_capacity_block = builder.create_block();
-
-            let big_enough =
-                builder
-                    .ins()
-                    .icmp(IntCC::UnsignedLessThanOrEqual, required_capacity, capacity);
-
-            builder.ins().brif(
-                big_enough,
-                sufficient_capacity_block,
-                &[],
-                insufficient_capacity_block,
-                &[],
-            );
-
-            {
-                builder.switch_to_block(insufficient_capacity_block);
-                builder.seal_block(insufficient_capacity_block);
-
-                emit_debug_println!(
-                    env,
-                    builder,
-                    "[ensure_capacity] need to increase capacity from {} to {}",
-                    capacity,
-                    required_capacity
-                );
-
-                if cfg!(debug_assertions) {
-                    // We must only re-allocate while there is no data in the buffer.
-                    let length = self.get_length(builder);
-                    emit_debug_assert_eq!(env, builder, length, zero);
-                }
-
-                let align = builder.ins().iconst(
-                    I64,
-                    std::mem::align_of::<wasmtime_continuations::types::payloads::DataEntries>()
-                        as i64,
-                );
-                let entry_size =
-                    std::mem::size_of::<wasmtime_continuations::types::payloads::DataEntries>();
-                let old_size = builder.ins().imul_imm(capacity, entry_size as i64);
-                let new_size = builder.ins().imul_imm(required_capacity, entry_size as i64);
-
-                let index = BuiltinFunctionIndex::tc_reallocate();
-                let sig = env.builtin_function_signatures.tc_reallocate(builder.func);
-
-                let ptr = self.get_data(builder);
-                let (_, new_data) = env.generate_builtin_call(
-                    builder,
-                    index,
-                    sig,
-                    vec![ptr, old_size, new_size, align],
-                );
-
-                self.set_capacity(builder, required_capacity);
-                self.set_data(builder, new_data);
-                self.set_length(builder, zero);
-                builder.ins().jump(sufficient_capacity_block, &[]);
-            }
-
-            builder.switch_to_block(sufficient_capacity_block);
-            builder.seal_block(sufficient_capacity_block);
-        }
-
-        /// Loads n entries from this Payloads object, where n is the length of
-        /// `load_types`, which also gives the types of the values to load.
-        /// Loading starts at index 0 of the Payloads object.
-        pub fn load_data_entries<'a>(
-            &self,
-            env: &mut crate::func_environ::FuncEnvironment<'a>,
-            builder: &mut FunctionBuilder,
-            load_types: &[ir::Type],
-        ) -> Vec<ir::Value> {
-            if cfg!(debug_assertions) {
-                let length = self.get_length(builder);
-                let load_count = builder.ins().iconst(I64, load_types.len() as i64);
-                emit_debug_assert_ule!(env, builder, load_count, length);
-            }
-
-            let memflags = ir::MemFlags::trusted();
-
-            let data_start_pointer = self.get_data(builder);
-            let mut values = vec![];
-            let mut offset = 0;
-            for valtype in load_types {
-                let val = builder
-                    .ins()
-                    .load(*valtype, memflags, data_start_pointer, offset);
-                values.push(val);
-                offset +=
-                    std::mem::size_of::<wasmtime_continuations::types::payloads::DataEntries>()
-                        as i32;
-            }
-            values
-        }
-
-        /// Stores the given `values` in this Payloads object, beginning at
-        /// index 0. This expects the Payloads object to be empty (i.e., current
-        /// length is 0), and to be of sufficient capacity to store |`values`|
-        /// entries.
-        pub fn store_data_entries<'a>(
-            &self,
-            env: &mut crate::func_environ::FuncEnvironment<'a>,
-            builder: &mut FunctionBuilder,
-            values: &[ir::Value],
-        ) {
-            let store_count = builder.ins().iconst(I64, values.len() as i64);
-
-            if cfg!(debug_assertions) {
-                let capacity = self.get_capacity(builder);
-                let length = self.get_length(builder);
-                let zero = builder.ins().iconst(I64, 0);
-                emit_debug_assert_ule!(env, builder, store_count, capacity);
-                emit_debug_assert_eq!(env, builder, length, zero);
-            }
-
-            let memflags = ir::MemFlags::trusted();
-
-            let data_start_pointer = self.get_data(builder);
-
-            let mut offset = 0;
-            for value in values {
-                builder
-                    .ins()
-                    .store(memflags, *value, data_start_pointer, offset);
-                offset +=
-                    std::mem::size_of::<wasmtime_continuations::types::payloads::DataEntries>()
-                        as i32;
-            }
-
-            self.set_length(builder, store_count);
-        }
-
-        pub fn clear(&self, builder: &mut FunctionBuilder) {
-            let zero = builder.ins().iconst(I64, 0);
-            self.set_length(builder, zero);
-        }
-
-        /// Silences some unused function warnings
-        #[allow(dead_code)]
-        pub fn dummy<'a>(
-            env: &mut crate::func_environ::FuncEnvironment<'a>,
-            builder: &mut FunctionBuilder,
-        ) {
-            let _index = BuiltinFunctionIndex::tc_allocate();
-            let _sig = env.builtin_function_signatures.tc_allocate(builder.func);
-            let _index = BuiltinFunctionIndex::tc_deallocate();
-            let _sig = env.builtin_function_signatures.tc_deallocate(builder.func);
-        }
-    }
-
-    impl VMContext {
-        pub fn new(address: ir::Value, pointer_type: ir::Type) -> VMContext {
-            VMContext {
-                address,
-                pointer_type,
-            }
-        }
-
-        pub fn get_active_continuation<'a>(
-            &self,
-            env: &mut crate::func_environ::FuncEnvironment<'a>,
-            builder: &mut FunctionBuilder,
-        ) -> ir::Value {
-            let base_addr = self.address;
-
-            let memflags = ir::MemFlags::trusted();
-            let offset = i32::try_from(env.offsets.vmctx_typed_continuations_store()).unwrap();
-            builder
-                .ins()
-                .load(self.pointer_type, memflags, base_addr, offset)
-        }
-
-        pub fn set_active_continuation<'a>(
-            &self,
-            env: &mut crate::func_environ::FuncEnvironment<'a>,
-            builder: &mut FunctionBuilder,
-            contobj: ir::Value,
-        ) {
-            let base_addr = self.address;
-
-            let memflags = ir::MemFlags::trusted();
-            let offset = i32::try_from(env.offsets.vmctx_typed_continuations_store()).unwrap();
-            builder.ins().store(memflags, contobj, base_addr, offset);
-        }
-    }
-}
-
-use typed_continuation_helpers as tc;
 
 impl<'module_environment> FuncEnvironment<'module_environment> {
     pub fn new(
@@ -952,11 +223,13 @@ impl<'module_environment> FuncEnvironment<'module_environment> {
         }
     }
 
-    fn pointer_type(&self) -> ir::Type {
+    // NOTE(frank-emrich) pub for use in crate::wasmfx::* modules
+    pub(crate) fn pointer_type(&self) -> ir::Type {
         self.isa.pointer_type()
     }
 
-    fn vmctx(&mut self, func: &mut Function) -> ir::GlobalValue {
+    // NOTE(frank-emrich) pub for use in crate::wasmfx::* modules
+    pub(crate) fn vmctx(&mut self, func: &mut Function) -> ir::GlobalValue {
         self.vmctx.unwrap_or_else(|| {
             let vmctx = func.create_global_value(ir::GlobalValueData::VMContext);
             if self.isa.flags().enable_pcc() {
@@ -1049,7 +322,8 @@ impl<'module_environment> FuncEnvironment<'module_environment> {
 
     /// Translates load of builtin function and returns a pair of values `vmctx`
     /// and address of the loaded function.
-    fn translate_load_builtin_function_address(
+    /// NOTE(frank-emrich) pub for use in crate::wasmfx::* modules
+    pub(crate) fn translate_load_builtin_function_address(
         &mut self,
         pos: &mut FuncCursor<'_>,
         callee_func_idx: BuiltinFunctionIndex,
@@ -1716,28 +990,6 @@ impl<'module_environment> FuncEnvironment<'module_environment> {
         builder.ins().call_indirect(sig, addr, &args);
         return vmctx;
     }
-}
-
-macro_rules! generate_builtin_call {
-    ($self : ident, $builder: ident, $builtin_name: ident, $args: expr) => {{
-        let index = BuiltinFunctionIndex::$builtin_name();
-        let sig = $self
-            .builtin_function_signatures
-            .$builtin_name(&mut $builder.func);
-        let args = $args.to_vec();
-        $self.generate_builtin_call($builder, index, sig, args)
-    }};
-}
-
-macro_rules! generate_builtin_call_no_return_val {
-    ($self : ident, $builder: ident, $builtin_name: ident, $args: expr) => {{
-        let index = BuiltinFunctionIndex::$builtin_name();
-        let sig = $self
-            .builtin_function_signatures
-            .$builtin_name(&mut $builder.func);
-        let args = $args.to_vec();
-        $self.generate_builtin_call_no_return_val($builder, index, sig, args)
-    }};
 }
 
 struct Call<'a, 'func, 'module_env> {
@@ -3455,16 +2707,7 @@ impl<'module_environment> cranelift_wasm::FuncEnvironment for FuncEnvironment<'m
         arg_types: &[WasmValType],
         return_types: &[WasmValType],
     ) -> WasmResult<ir::Value> {
-        if cfg!(feature = "typed_continuations_baseline_implementation") {
-            return self.translate_cont_new_baseline(builder, state, func, arg_types, return_types);
-        }
-        let nargs = builder.ins().iconst(I64, arg_types.len() as i64);
-        let nreturns = builder.ins().iconst(I64, return_types.len() as i64);
-
-        let (_vmctx, contobj) =
-            generate_builtin_call!(self, builder, tc_cont_new, [func, nargs, nreturns]);
-
-        Ok(contobj)
+        wasmfx_impl::translate_cont_new(self, builder, state, func, arg_types, return_types)
     }
 
     // TODO(dhil): Currently, this function invokes
@@ -3479,246 +2722,7 @@ impl<'module_environment> cranelift_wasm::FuncEnvironment for FuncEnvironment<'m
         resume_args: &[ir::Value],
         resumetable: &[(u32, ir::Block)],
     ) -> Vec<ir::Value> {
-        if cfg!(feature = "typed_continuations_baseline_implementation") {
-            return self.translate_resume_baseline(
-                builder,
-                type_index,
-                contref,
-                resume_args,
-                resumetable,
-            );
-        }
-        let resume_block = builder.create_block();
-        let return_block = builder.create_block();
-        let suspend_block = builder.create_block();
-        let switch_block = builder.create_block();
-        let forwarding_block = builder.create_block();
-
-        // Preamble: Part of previously active block
-        {
-            let original_contobj = self.typed_continuations_cont_ref_get_cont_obj(builder, contref);
-
-            if resume_args.len() > 0 {
-                // We store the arguments in the continuation object to be resumed.
-                let count = builder.ins().iconst(I64, resume_args.len() as i64);
-                self.typed_continuations_store_resume_args(
-                    builder,
-                    resume_args,
-                    count,
-                    original_contobj,
-                );
-            }
-
-            // Make the currently running continuation the parent of the one we are about to resume.
-            let original_running_contobj =
-                self.typed_continuations_load_continuation_object(builder);
-            self.typed_continuations_store_parent(
-                builder,
-                original_contobj,
-                original_running_contobj,
-            );
-
-            builder.ins().jump(resume_block, &[original_contobj]);
-        }
-
-        // Resume block: actually resume the fiber corresponding to the
-        // continuation object given as a parameter to the block. This
-        // parameterisation is necessary to enable forwarding, requiring us
-        // to resume objects other than `original_contobj`.
-        // We make the continuation object that was actually resumed available via
-        // `resumed_contobj`, so that subsequent blocks can refer to it.
-        let (tag, resumed_contobj) = {
-            builder.switch_to_block(resume_block);
-            builder.append_block_param(resume_block, self.pointer_type());
-
-            // The continuation object to actually call resume on.
-            let resume_contobj = builder.block_params(resume_block)[0];
-
-            let vmctx = self.vmctx(builder.func);
-            let vmctx = builder.ins().global_value(self.pointer_type(), vmctx);
-
-            // We mark `resume_contobj` as the currently running one
-            let vmctx = tc::VMContext::new(vmctx, self.pointer_type());
-            vmctx.set_active_continuation(self, builder, resume_contobj);
-
-            // We mark `resume_contobj` to be invoked
-            let co = tc::ContinuationObject::new(resume_contobj, self.pointer_type());
-            co.set_state(builder, wasmtime_continuations::State::Invoked);
-
-            let (_vmctx, result) =
-                generate_builtin_call!(self, builder, tc_resume, [resume_contobj]);
-
-            // Now the parent contobj is active
-            let parent_contobj = self.typed_continuations_load_parent(builder, resume_contobj);
-            vmctx.set_active_continuation(self, builder, parent_contobj);
-
-            // The result encodes whether the return happens via ordinary
-            // means or via a suspend. If the high bit is set, then it is
-            // interpreted as the return happened via a suspend, and the
-            // remainder of the integer is to be interpreted as the index
-            // of the control tag that was supplied to the suspend.
-            let signal_mask = 0xf000_0000;
-            let inverse_signal_mask = 0x0fff_ffff;
-            let signal = builder.ins().band_imm(result, signal_mask);
-            let tag = builder.ins().band_imm(result, inverse_signal_mask);
-
-            // Description of results:
-            // * The `base_addr` is the base address of VM context.
-            // * The `signal` is an encoded boolean indicating whether
-            // the `resume` returned ordinarily or via a suspend
-            // instruction.
-            // * The `tag` is the index of the control tag supplied to
-            // suspend (only valid if `signal` is 1).
-
-            // Test the signal bit.
-            let is_zero = builder.ins().icmp_imm(IntCC::Equal, signal, 0);
-
-            // Jump to the return block if the signal is 0, otherwise
-            // jump to the suspend block.
-            builder
-                .ins()
-                .brif(is_zero, return_block, &[], suspend_block, &[]);
-
-            // We do not seal this block, yet, because the effect forwarding block has a back edge to it
-            (tag, resume_contobj)
-        };
-
-        // Suspend block.
-        {
-            builder.switch_to_block(suspend_block);
-            builder.seal_block(suspend_block);
-
-            // We need to terminate this block before being allowed to switch to another one
-            builder.ins().jump(switch_block, &[]);
-        };
-
-        // Now, construct blocks for the three continuations:
-        // 1) `resume` returned normally.
-        // 2) `resume` returned via a suspend.
-        // 3) `resume` is forwarding
-
-        // Strategy:
-        //
-        // Translate each each `(tag, label)` pair in the resume table
-        // to a switch-case of the form "case tag: br label". NOTE:
-        // `tag` may appear multiple times in resume table, only the
-        // first appearance should be processed as it shadows the
-        // subsequent entries.  The switching logic then ensures that
-        // we jump to the block handling the corresponding tag.
-        //
-        // The fallback/default case performs effect forwarding (TODO).
-        //
-        // First, initialise the switch structure.
-        let mut switch = Switch::new();
-        // Second, we consume the resume table entry-wise.
-        let mut case_blocks = vec![];
-        let mut tag_seen = std::collections::HashSet::new(); // Used to keep track of tags
-        for &(tag, target_block) in resumetable {
-            // Skip if this `tag` has been seen previously.
-            if !tag_seen.insert(tag) {
-                continue;
-            }
-            let case = builder.create_block();
-            switch.set_entry(tag as u128, case);
-            builder.switch_to_block(case);
-
-            // Load and push arguments.
-            let param_types = self.tag_params(tag);
-            let param_types: Vec<ir::Type> = param_types
-                .iter()
-                .map(|wty| crate::value_type(self.isa, *wty))
-                .collect();
-            let mut args = self.typed_continuations_load_payloads(builder, &param_types);
-
-            // We have an actual handling block for this tag, rather than just forwarding.
-            // Detatch the continuation object by setting its parent to NULL.
-            let pointer_type = self.pointer_type();
-            let null = builder.ins().iconst(pointer_type, 0);
-            self.typed_continuations_store_parent(builder, resumed_contobj, null);
-
-            // Create and push the continuation reference. We only create
-            // them here because we don't need them when forwarding.
-            let contref = self.typed_continuations_new_cont_ref(builder, resumed_contobj);
-
-            args.push(contref);
-
-            // Now jump to the actual user-defined block handling
-            // this tag, as given by the resumetable.
-            builder.ins().jump(target_block, &args);
-            case_blocks.push(case);
-        }
-
-        // Note that at this point we haven't actually emitted any
-        // code for the switching logic itself, but only filled
-        // the Switch structure and created the blocks it jumps
-        // to.
-
-        // Forwarding block: Default case for the switching logic on the
-        // tag. Used when the (resume ...) clause we currently translate
-        // does not have a matching (tag ...) entry.
-        {
-            builder.switch_to_block(forwarding_block);
-
-            let parent_contobj = self.typed_continuations_load_parent(builder, resumed_contobj);
-
-            builder
-                .ins()
-                .trapz(parent_contobj, ir::TrapCode::UnhandledTag);
-
-            // We suspend, thus deferring handling to the parent.
-            // We do nothing about tag *parameters*, these remain unchanged within the
-            // payload buffer associated with the whole VMContext.
-            generate_builtin_call_no_return_val!(self, builder, tc_suspend, [tag]);
-
-            // "Tag return values" (i.e., values provided by cont.bind or
-            // resume to the continuation) are actually stored in
-            // continuation objects, and we need to move them down the chain
-            // back to the continuation object where we originally
-            // suspended.
-            self.typed_continuations_forward_tag_return_values(
-                builder,
-                parent_contobj,
-                resumed_contobj,
-            );
-
-            builder.ins().jump(resume_block, &[resumed_contobj]);
-            builder.seal_block(resume_block);
-        }
-
-        // Switch block: actual switching logic is emitted here.
-        {
-            builder.switch_to_block(switch_block);
-            switch.emit(builder, tag, forwarding_block);
-            builder.seal_block(switch_block);
-            builder.seal_block(forwarding_block);
-
-            // We can only seal the blocks we generated for each
-            // tag now, after switch.emit ran.
-            for case_block in case_blocks {
-                builder.seal_block(case_block);
-            }
-        }
-
-        // Return block: Jumped to by resume block if continuation returned normally.
-        {
-            builder.switch_to_block(return_block);
-            builder.seal_block(return_block);
-
-            let co = tc::ContinuationObject::new(resumed_contobj, self.pointer_type());
-            co.set_state(builder, wasmtime_continuations::State::Returned);
-
-            // Load and push the results.
-            let returns = self.continuation_returns(type_index).to_vec();
-            let values =
-                self.typed_continuations_load_return_values(builder, &returns, resumed_contobj);
-
-            // The continuation has returned and all `ContinuationReferences`
-            // to it should have been be invalidated. We may safely deallocate
-            // it.
-            self.typed_continuations_drop_cont_obj(builder, resumed_contobj);
-
-            return values;
-        }
+        wasmfx_impl::translate_resume(self, builder, type_index, contref, resume_args, resumetable)
     }
 
     fn translate_resume_throw(
@@ -3736,11 +2740,7 @@ impl<'module_environment> cranelift_wasm::FuncEnvironment for FuncEnvironment<'m
         builder: &mut FunctionBuilder,
         tag_index: ir::Value,
     ) -> ir::Value {
-        if cfg!(feature = "typed_continuations_baseline_implementation") {
-            return self.translate_suspend_baseline(builder, tag_index);
-        }
-        // Returns the vmctx
-        return generate_builtin_call_no_return_val!(self, builder, tc_suspend, [tag_index]);
+        wasmfx_impl::translate_suspend(self, builder, tag_index)
     }
 
     fn continuation_arguments(&self, index: u32) -> &[WasmValType] {
@@ -3763,139 +2763,13 @@ impl<'module_environment> cranelift_wasm::FuncEnvironment for FuncEnvironment<'m
         self.types[idx].returns()
     }
 
-    fn typed_continuations_load_payloads(
-        &mut self,
-        builder: &mut FunctionBuilder,
-        valtypes: &[ir::Type],
-    ) -> Vec<ir::Value> {
-        let mut values = vec![];
-
-        if valtypes.len() > 0 {
-            if cfg!(feature = "typed_continuations_baseline_implementation") {
-                // Retrieve the pointer to the suspend buffer.
-                let index = BuiltinFunctionIndex::tc_baseline_get_payloads_ptr();
-                let sig = self
-                    .builtin_function_signatures
-                    .tc_baseline_get_payloads_ptr(builder.func);
-                let nargs = builder.ins().iconst(I64, 0);
-                let (_vmctx, payloads_ptr) =
-                    self.generate_builtin_call(builder, index, sig, vec![nargs]);
-                // Load payloads.
-                let memflags = ir::MemFlags::trusted();
-                let mut offset = 0;
-                for valtype in valtypes {
-                    let val = builder.ins().load(*valtype, memflags, payloads_ptr, offset);
-                    values.push(val);
-                    offset += self.offsets.ptr.maximum_value_size() as i32;
-                }
-                // Clear the payloads buffer
-                let index = BuiltinFunctionIndex::tc_baseline_clear_payloads();
-                let sig = self
-                    .builtin_function_signatures
-                    .tc_baseline_clear_payloads(builder.func);
-                self.generate_builtin_call_no_return_val(builder, index, sig, vec![]);
-            } else {
-                let vmctx = self.vmctx(builder.cursor().func);
-                let vmctx = builder.ins().global_value(self.pointer_type(), vmctx);
-                let vmctx_payloads = tc::Payloads::new(
-                    vmctx,
-                    self.offsets.vmctx_typed_continuations_payloads() as i32,
-                    self.pointer_type(),
-                );
-
-                values = vmctx_payloads.load_data_entries(self, builder, valtypes);
-
-                // In theory, we way want to deallocate the buffer instead of just
-                // clearing it if its size is above a certain threshold. That would
-                // avoid keeping a large object unnecessarily long.
-                vmctx_payloads.clear(builder);
-            }
-        }
-        values
-    }
-
     fn typed_continuations_load_tag_return_values(
         &mut self,
         builder: &mut FunctionBuilder,
         contobj: ir::Value,
         valtypes: &[WasmValType],
     ) -> Vec<ir::Value> {
-        let memflags = ir::MemFlags::trusted();
-        let mut values = vec![];
-
-        if valtypes.len() > 0 {
-            if cfg!(feature = "typed_continuations_baseline_implementation") {
-                // Retrieve the pointer to the arguments' buffer.
-                let index = BuiltinFunctionIndex::tc_baseline_continuation_arguments_ptr();
-                let sig = self
-                    .builtin_function_signatures
-                    .tc_baseline_continuation_arguments_ptr(builder.func);
-                let nargs = builder.ins().iconst(I64, 0);
-                let (_vmctx, args_ptr) =
-                    self.generate_builtin_call(builder, index, sig, vec![contobj, nargs]);
-                // Load arguments.
-                let mut args = vec![];
-                let mut offset = 0;
-                let memflags = ir::MemFlags::trusted();
-                for valtype in valtypes {
-                    let val = builder.ins().load(
-                        super::value_type(self.isa, *valtype),
-                        memflags,
-                        args_ptr,
-                        offset,
-                    );
-                    args.push(val);
-                    offset += self.offsets.ptr.maximum_value_size() as i32;
-                }
-                debug_assert!(valtypes.len() == args.len());
-
-                // Clear the arguments buffer
-                let index = BuiltinFunctionIndex::tc_baseline_clear_arguments();
-                let sig = self
-                    .builtin_function_signatures
-                    .tc_baseline_clear_arguments(builder.func);
-                self.generate_builtin_call_no_return_val(builder, index, sig, vec![contobj]);
-
-                return args;
-            } else {
-                let co = tc::ContinuationObject::new(contobj, self.pointer_type());
-                let tag_return_values = co.tag_return_values();
-
-                let payload_ptr = tag_return_values.get_data(builder);
-
-                let mut offset = 0;
-                for valtype in valtypes {
-                    let val = builder.ins().load(
-                        super::value_type(self.isa, *valtype),
-                        memflags,
-                        payload_ptr,
-                        offset,
-                    );
-                    values.push(val);
-                    offset += self.offsets.ptr.maximum_value_size() as i32;
-                }
-
-                // In theory, we way want to deallocate the buffer instead of just
-                // clearing it if its size is above a certain threshold. That would
-                // avoid keeping a large object unnecessarily long.
-                tag_return_values.clear(builder);
-            }
-        }
-        values
-    }
-
-    fn typed_continuations_forward_tag_return_values(
-        &mut self,
-        builder: &mut FunctionBuilder,
-        parent_contobj: ir::Value,
-        child_contobj: ir::Value,
-    ) {
-        generate_builtin_call_no_return_val!(
-            self,
-            builder,
-            tc_cont_obj_forward_tag_return_values_buffer,
-            [parent_contobj, child_contobj]
-        );
+        wasmfx_impl::typed_continuations_load_tag_return_values(self, builder, contobj, valtypes)
     }
 
     /// TODO
@@ -3904,14 +2778,7 @@ impl<'module_environment> cranelift_wasm::FuncEnvironment for FuncEnvironment<'m
         builder: &mut FunctionBuilder,
         contref: ir::Value,
     ) -> ir::Value {
-        if cfg!(feature = "unsafe_disable_continuation_linearity_check") {
-            // The "contref" is a contobj already
-            return contref;
-        } else {
-            let (_vmctx, contobj) =
-                generate_builtin_call!(self, builder, tc_cont_ref_get_cont_obj, [contref]);
-            return contobj;
-        }
+        wasmfx_impl::typed_continuations_cont_ref_get_cont_obj(self, builder, contref)
     }
 
     /// TODO
@@ -3922,78 +2789,13 @@ impl<'module_environment> cranelift_wasm::FuncEnvironment for FuncEnvironment<'m
         remaining_arg_count: ir::Value,
         contobj: ir::Value,
     ) {
-        if values.len() > 0 {
-            if cfg!(feature = "typed_continuations_baseline_implementation") {
-                // Retrieve the pointer to the arguments buffer.
-                let index = BuiltinFunctionIndex::tc_baseline_continuation_arguments_ptr();
-                let sig = self
-                    .builtin_function_signatures
-                    .tc_baseline_continuation_arguments_ptr(builder.func);
-                let nargs = builder.ins().iconst(I64, values.len() as i64);
-                let (_vmctx, args_ptr) =
-                    self.generate_builtin_call(builder, index, sig, vec![contobj, nargs]);
-                // Store arguments.
-                let memflags = ir::MemFlags::trusted();
-                let mut offset = 0;
-                for arg in values {
-                    builder.ins().store(memflags, *arg, args_ptr, offset);
-                    offset += self.offsets.ptr.maximum_value_size() as i32;
-                }
-            } else {
-                let use_args_block = builder.create_block();
-                let use_payloads_block = builder.create_block();
-                let store_data_block = builder.create_block();
-                builder.append_block_param(store_data_block, self.pointer_type());
-
-                let co = tc::ContinuationObject::new(contobj, self.pointer_type());
-                let is_invoked = co.is_invoked(builder);
-                builder
-                    .ins()
-                    .brif(is_invoked, use_payloads_block, &[], use_args_block, &[]);
-
-                {
-                    builder.switch_to_block(use_args_block);
-                    builder.seal_block(use_args_block);
-
-                    let args = co.args();
-                    let ptr = args.occupy_next_slots(self, builder, values.len() as i32);
-
-                    builder.ins().jump(store_data_block, &[ptr]);
-                }
-
-                {
-                    builder.switch_to_block(use_payloads_block);
-                    builder.seal_block(use_payloads_block);
-
-                    let tag_return_values = co.tag_return_values();
-
-                    // Unlike for the args buffer (where we know the maximum
-                    // required capacity at the time of creation of the
-                    // ContinuationObject), tag return buffers are re-used and may
-                    // be too small.
-                    tag_return_values.ensure_capacity(self, builder, remaining_arg_count);
-
-                    let ptr =
-                        tag_return_values.occupy_next_slots(self, builder, values.len() as i32);
-                    builder.ins().jump(store_data_block, &[ptr]);
-                }
-
-                {
-                    builder.switch_to_block(store_data_block);
-                    builder.seal_block(store_data_block);
-
-                    let ptr = builder.block_params(store_data_block)[0];
-
-                    // Store the values.
-                    let memflags = ir::MemFlags::trusted();
-                    let mut offset = 0;
-                    for value in values {
-                        builder.ins().store(memflags, *value, ptr, offset);
-                        offset += self.offsets.ptr.maximum_value_size() as i32;
-                    }
-                }
-            }
-        }
+        wasmfx_impl::typed_continuations_store_resume_args(
+            self,
+            builder,
+            values,
+            remaining_arg_count,
+            contobj,
+        )
     }
 
     //TODO(frank-emrich) Consider removing `valtypes` argument, as values are inherently typed
@@ -4003,61 +2805,14 @@ impl<'module_environment> cranelift_wasm::FuncEnvironment for FuncEnvironment<'m
         valtypes: &[WasmValType],
         values: &[ir::Value],
     ) {
-        assert_eq!(values.len(), valtypes.len());
-        if valtypes.len() > 0 {
-            if cfg!(feature = "typed_continuations_baseline_implementation") {
-                // Retrieve the pointer to the payloads buffer.
-                let index = BuiltinFunctionIndex::tc_baseline_get_payloads_ptr();
-                let sig = self
-                    .builtin_function_signatures
-                    .tc_baseline_get_payloads_ptr(builder.func);
-                let nargs = builder.ins().iconst(I64, values.len() as i64);
-                let (_vmctx, payloads_ptr) =
-                    self.generate_builtin_call(builder, index, sig, vec![nargs]);
-                // Store arguments.
-                let memflags = ir::MemFlags::trusted();
-                let mut offset = 0;
-                for arg in values {
-                    builder.ins().store(memflags, *arg, payloads_ptr, offset);
-                    offset += self.offsets.ptr.maximum_value_size() as i32;
-                }
-            } else {
-                let vmctx = self.vmctx(builder.cursor().func);
-                let vmctx = builder.ins().global_value(self.pointer_type(), vmctx);
-                let payloads = tc::Payloads::new(
-                    vmctx,
-                    self.offsets.vmctx_typed_continuations_payloads() as i32,
-                    self.pointer_type(),
-                );
-
-                let nargs = builder.ins().iconst(I64, values.len() as i64);
-                payloads.ensure_capacity(self, builder, nargs);
-
-                payloads.store_data_entries(self, builder, values);
-            }
-        }
+        wasmfx_impl::typed_continuations_store_payloads(self, builder, valtypes, values)
     }
 
     fn typed_continuations_load_continuation_object(
         &mut self,
         builder: &mut FunctionBuilder,
     ) -> ir::Value {
-        if cfg!(feature = "typed_continuations_baseline_implementation") {
-            let index = BuiltinFunctionIndex::tc_baseline_get_current_continuation();
-            let sig = self
-                .builtin_function_signatures
-                .tc_baseline_get_current_continuation(builder.func);
-            let (_vmctx, contobj) = self.generate_builtin_call(builder, index, sig, vec![]);
-            return contobj;
-        } else {
-            let pointer_type = self.pointer_type();
-            let vmctx = self.vmctx(builder.cursor().func);
-            let base_addr = builder.ins().global_value(pointer_type, vmctx);
-
-            let vmctx = tc::VMContext::new(base_addr, pointer_type);
-
-            vmctx.get_active_continuation(self, builder)
-        }
+        wasmfx_impl::typed_continuations_load_continuation_object(self, builder)
     }
 
     fn typed_continuations_new_cont_ref(
@@ -4065,404 +2820,8 @@ impl<'module_environment> cranelift_wasm::FuncEnvironment for FuncEnvironment<'m
         builder: &mut FunctionBuilder,
         contobj_addr: ir::Value,
     ) -> ir::Value {
-        if cfg!(feature = "unsafe_disable_continuation_linearity_check") {
-            return contobj_addr;
-        } else {
-            let (_vmctx, contref) =
-                generate_builtin_call!(self, builder, tc_new_cont_ref, [contobj_addr]);
-            return contref;
-        }
+        wasmfx_impl::typed_continuations_new_cont_ref(self, builder, contobj_addr)
     }
-
-    fn typed_continuations_load_parent(
-        &mut self,
-        builder: &mut FunctionBuilder,
-        contobj: ir::Value,
-    ) -> ir::Value {
-        let offset = wasmtime_continuations::offsets::continuation_object::PARENT;
-        let memflags = ir::MemFlags::trusted();
-
-        builder
-            .ins()
-            .load(self.pointer_type(), memflags, contobj, offset)
-    }
-
-    fn typed_continuations_store_parent(
-        &mut self,
-        builder: &mut FunctionBuilder,
-        contobj: ir::Value,
-        new_parent: ir::Value,
-    ) {
-        let offset = wasmtime_continuations::offsets::continuation_object::PARENT;
-        let memflags = ir::MemFlags::trusted();
-
-        builder.ins().store(memflags, new_parent, contobj, offset);
-    }
-
-    /// TODO
-    fn typed_continuations_drop_cont_obj(
-        &mut self,
-        builder: &mut FunctionBuilder,
-        contobj: ir::Value,
-    ) {
-        generate_builtin_call_no_return_val!(self, builder, tc_drop_cont_obj, [contobj]);
-    }
-
-    fn typed_continuations_load_return_values(
-        &mut self,
-        builder: &mut FunctionBuilder,
-        valtypes: &[WasmValType],
-        contobj: ir::Value,
-    ) -> std::vec::Vec<ir::Value> {
-        let co = tc::ContinuationObject::new(contobj, self.pointer_type());
-        let mut values = vec![];
-
-        if valtypes.len() > 0 {
-            let result_buffer_addr = co.get_results(self, builder);
-
-            let mut offset = 0;
-            let memflags = ir::MemFlags::trusted();
-            for valtype in valtypes {
-                let val = builder.ins().load(
-                    super::value_type(self.isa, *valtype),
-                    memflags,
-                    result_buffer_addr,
-                    offset,
-                );
-                values.push(val);
-                offset += self.offsets.ptr.maximum_value_size() as i32;
-            }
-        }
-        return values;
-    }
-
-    //
-    // Typed continuation baseline implementations of cont_new, resume, suspend.
-    //
-    fn translate_cont_new_baseline(
-        &mut self,
-        builder: &mut FunctionBuilder,
-        _state: &FuncTranslationState,
-        func: ir::Value,
-        arg_types: &[WasmValType],
-        return_types: &[WasmValType],
-    ) -> WasmResult<ir::Value> {
-        // Load the builtin continuation allocation function.
-        let func_sig = self
-            .builtin_function_signatures
-            .tc_baseline_cont_new(builder.func);
-        let (vmctx, func_addr) = self.translate_load_builtin_function_address(
-            &mut builder.cursor(),
-            BuiltinFunctionIndex::tc_baseline_cont_new(),
-        );
-        // Call the builtin.
-        let nargs = builder.ins().iconst(I64, arg_types.len() as i64);
-        let nreturns = builder.ins().iconst(I64, return_types.len() as i64);
-        let call_inst =
-            builder
-                .ins()
-                .call_indirect(func_sig, func_addr, &[vmctx, func, nargs, nreturns]);
-        // Unwrap and take ownership of the result.
-        assert!(builder.func.dfg.has_results(call_inst));
-        let contref = *builder.func.dfg.inst_results(call_inst).first().unwrap();
-
-        Ok(contref)
-    }
-
-    fn translate_resume_baseline(
-        &mut self,
-        builder: &mut FunctionBuilder,
-        type_index: u32,
-        resumee_ref: ir::Value,
-        resume_args: &[ir::Value],
-        resumetable: &[(u32, ir::Block)],
-    ) -> Vec<ir::Value> {
-        // The resume instruction is by far the most involved
-        // instruction to compile as it is responsible for both
-        // continuation application and effect dispatch.
-        //
-        // We store the continuation arguments, continuation return
-        // values, and suspension payloads on objects accessible
-        // inside the libcall context.
-        //
-        // Here we translate a resume instruction into several basic
-        // blocks as follows:
-        //
-        //        prelude_block
-        //              |
-        //              |
-        //        resume_block <--------
-        //        |          |          \
-        //        |          |          |
-        // return_block   handle_block  |
-        //                     |        |
-        //                     |        |
-        //                forward_block /
-        //
-        // * prelude_block pushes the continuation arguments onto the
-        //   buffer in the libcall context.
-        // * resume_block continues a given `resumee_ref`. It jumps to
-        //   return_block on ordinary return and to `handle_block` on
-        //   a suspension
-        // * return_block reads the return values from the libcall
-        //   context.
-        // * handle_block dispatches on a tag provided by the
-        //   resume_block to an associated user-defined block. If
-        //   there is no suitable user-defined block, then it jumps to
-        //   the forward_block.
-        // * forward_block dispatches the handling of a given tag to
-        //   the ambient context. Once control returns it jumps to the
-        //   resume_block to continue continuation at the suspension
-        //   site.
-        let resume_block = builder.create_block();
-        let return_block = builder.create_block();
-        let handle_block = builder.create_block();
-        let forwarding_block = builder.create_block();
-
-        // Prelude: Push the continuation arguments.
-        {
-            let resumee_fiber =
-                self.typed_continuations_cont_ref_get_cont_obj(builder, resumee_ref);
-            if resume_args.len() > 0 {
-                let nargs = builder.ins().iconst(I64, resume_args.len() as i64);
-
-                // Load the arguments pointer
-                let func_sig = self
-                    .builtin_function_signatures
-                    .tc_baseline_continuation_arguments_ptr(builder.func);
-                let (vmctx, func_addr) = self.translate_load_builtin_function_address(
-                    &mut builder.cursor(),
-                    BuiltinFunctionIndex::tc_baseline_continuation_arguments_ptr(),
-                );
-                // Call the builtin.
-                let call_inst = builder.ins().call_indirect(
-                    func_sig,
-                    func_addr,
-                    &[vmctx, resumee_fiber, nargs],
-                );
-                // Unwrap and take ownership of the result.
-                debug_assert!(builder.func.dfg.has_results(call_inst));
-                let args_ptr = *builder.func.dfg.inst_results(call_inst).first().unwrap();
-
-                // Append arguments.
-                let memflags = ir::MemFlags::trusted();
-                let mut offset = 0;
-                for arg in resume_args {
-                    builder.ins().store(memflags, *arg, args_ptr, offset);
-                    offset += self.offsets.ptr.maximum_value_size() as i32;
-                }
-            }
-
-            builder.ins().jump(resume_block, &[resumee_fiber]);
-        }
-
-        // Resume block: here we continue the (suspended) resumee.
-        let (tag, resumee_fiber) = {
-            builder.switch_to_block(resume_block);
-            builder.append_block_param(resume_block, self.pointer_type());
-
-            // The raw continuation fiber
-            let resumee_fiber = builder.block_params(resume_block)[0];
-
-            // Load the builtin continuation resume function.
-            let func_sig = self
-                .builtin_function_signatures
-                .tc_baseline_resume(builder.func);
-            let (vmctx, func_addr) = self.translate_load_builtin_function_address(
-                &mut builder.cursor(),
-                BuiltinFunctionIndex::tc_baseline_resume(),
-            );
-
-            // Continue the resumee.
-            let call_inst =
-                builder
-                    .ins()
-                    .call_indirect(func_sig, func_addr, &[vmctx, resumee_fiber]);
-
-            // Unwrap and take ownership of the result.
-            debug_assert!(builder.func.dfg.has_results(call_inst));
-            let result = *builder.func.dfg.inst_results(call_inst).first().unwrap();
-
-            // The result encodes whether the return happens via ordinary
-            // means or via a suspend. If the high bit is set, then it is
-            // interpreted as the return happened via a suspend, and the
-            // remainder of the integer is to be interpreted as the index
-            // of the control tag that was supplied to the suspend.
-            let signal_mask = 0xf000_0000;
-            let inverse_signal_mask = 0x0fff_ffff;
-            let signal = builder.ins().band_imm(result, signal_mask);
-            let tag = builder.ins().band_imm(result, inverse_signal_mask);
-
-            // Test the signal bit.
-            let is_zero = builder.ins().icmp_imm(IntCC::Equal, signal, 0);
-
-            // Jump to the return block if the signal is 0, otherwise
-            // jump to the suspend block.
-            builder
-                .ins()
-                .brif(is_zero, return_block, &[], handle_block, &[]);
-
-            // We do not seal this block, yet, because the effect forwarding block has a back edge to it
-            (tag, resumee_fiber)
-        };
-
-        // Now we construct the handling block.
-        //
-        // Strategy:
-        //
-        // We encode the resume table as a sparse switch table, where
-        // each `(tag, label)` pair in the resume table translates to
-        // a "case tag: br label" in the switch table.
-        //
-        // The first occurs of a particular tag `t` shadows every
-        // other occurrence of the same tag `t` in the resumetable,
-        // meaning we should only emit code for the first such `t`.
-        //
-        // Effect forwarding is handled by the default/fallthrough
-        // case of the switch.
-
-        // First, initialise the switch structure.
-        let mut switch = Switch::new();
-        // Second, we consume the resume table entry-wise.
-        let mut case_blocks = vec![];
-        let mut tag_seen = std::collections::HashSet::new(); // Used to keep track of tags
-        for &(tag, label) in resumetable {
-            // Skip if this `tag` has been seen previously.
-            if !tag_seen.insert(tag) {
-                continue;
-            }
-            let case = builder.create_block();
-            switch.set_entry(tag as u128, case);
-            builder.switch_to_block(case);
-
-            // Load and push payloads.
-            let param_types: Vec<ir::Type> = self
-                .tag_params(tag)
-                .iter()
-                .map(|wty| crate::value_type(self.isa, *wty))
-                .collect();
-            // NOTE(dhil): For the baseline, `load_payloads` actually
-            // moves the payloads, i.e. consumes the payloads buffer
-            // entirely.
-            let mut args = self.typed_continuations_load_payloads(builder, &param_types);
-
-            // Create and push the continuation reference.
-            let resumee_ref = self.typed_continuations_new_cont_ref(builder, resumee_fiber);
-            args.push(resumee_ref);
-
-            // Finally, emit the jump to `label`.
-            builder.ins().jump(label, &args);
-            case_blocks.push(case);
-        }
-
-        // Forwarding block.
-        {
-            builder.switch_to_block(forwarding_block);
-
-            // Load the builtin forwarding function.
-            let func_sig = self
-                .builtin_function_signatures
-                .tc_baseline_forward(builder.func);
-            let (vmctx, func_addr) = self.translate_load_builtin_function_address(
-                &mut builder.cursor(),
-                BuiltinFunctionIndex::tc_baseline_forward(),
-            );
-
-            // Continue the resumee.
-            let call_inst =
-                builder
-                    .ins()
-                    .call_indirect(func_sig, func_addr, &[vmctx, tag, resumee_fiber]);
-            // Check the result.
-            debug_assert!(!builder.func.dfg.has_results(call_inst));
-
-            builder.ins().jump(resume_block, &[resumee_fiber]);
-            builder.seal_block(resume_block);
-        }
-
-        // Emit the switch.
-        {
-            builder.switch_to_block(handle_block);
-            switch.emit(builder, tag, forwarding_block);
-            builder.seal_block(handle_block);
-            builder.seal_block(forwarding_block);
-
-            for case_block in case_blocks {
-                builder.seal_block(case_block);
-            }
-        }
-
-        // Return block.
-        {
-            builder.switch_to_block(return_block);
-            builder.seal_block(return_block);
-
-            // Load the values pointer.
-            let func_sig = self
-                .builtin_function_signatures
-                .tc_baseline_continuation_values_ptr(builder.func);
-            let (vmctx, func_addr) = self.translate_load_builtin_function_address(
-                &mut builder.cursor(),
-                BuiltinFunctionIndex::tc_baseline_continuation_values_ptr(),
-            );
-            // Call the builtin.
-            let call_inst =
-                builder
-                    .ins()
-                    .call_indirect(func_sig, func_addr, &[vmctx, resumee_fiber]);
-            // Unwrap and take ownership of the result.
-            debug_assert!(builder.func.dfg.has_results(call_inst));
-            let vals_ptr = *builder.func.dfg.inst_results(call_inst).first().unwrap();
-
-            // Load and push the return values.
-            let returns = self.continuation_returns(type_index);
-            let mut values = vec![];
-            let mut offset = 0;
-            let memflags = ir::MemFlags::trusted();
-            for valtype in returns {
-                let val = builder.ins().load(
-                    super::value_type(self.isa, *valtype),
-                    memflags,
-                    vals_ptr,
-                    offset,
-                );
-                values.push(val);
-                offset += self.offsets.ptr.maximum_value_size() as i32;
-            }
-
-            // Free the continuation.
-            let func_sig = self
-                .builtin_function_signatures
-                .tc_baseline_drop_continuation_reference(builder.func);
-            let (vmctx, func_addr) = self.translate_load_builtin_function_address(
-                &mut builder.cursor(),
-                BuiltinFunctionIndex::tc_baseline_drop_continuation_reference(),
-            );
-            // Call the builtin.
-            let call_inst =
-                builder
-                    .ins()
-                    .call_indirect(func_sig, func_addr, &[vmctx, resumee_fiber]);
-            // Check the result.
-            debug_assert!(!builder.func.dfg.has_results(call_inst));
-
-            return values;
-        }
-    }
-
-    fn translate_suspend_baseline(
-        &mut self,
-        builder: &mut FunctionBuilder,
-        tag_index: ir::Value,
-    ) -> ir::Value {
-        let index = BuiltinFunctionIndex::tc_baseline_suspend();
-        let sig = self
-            .builtin_function_signatures
-            .tc_baseline_suspend(builder.func);
-        // Returns the vmctx
-        return self.generate_builtin_call_no_return_val(builder, index, sig, vec![tag_index]);
-    }
-    // End of baseline implementations.
 
     fn use_x86_blendv_for_relaxed_laneselect(&self, ty: Type) -> bool {
         self.isa.has_x86_blendv_lowering(ty)
